cloud_function/gcs_pd_bq/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
